refactor(RouteSegment): Remove distance and path from RouteSegment
- Updating the `RouteSegment` data class to remove `distanceMeter` and `path`.
- Removing the `getDistanceKm()` method from `RouteSegment`.
- Adjusting `TripsRepositoryFirestore` to no longer map `distanceMeter` and `path` from Firestore documents.
- Updating the `TripScheduler` to no longer set these properties when creating `RouteSegment` instances.
- Modifying all related tests (`TripTest`, `RouteSegmentTest`, `TripsRepositoryMockTest`) to reflect the updated `RouteSegment` structure.
- Adding `CAR` to the `TransportMode` enum.
