+++ app/src/main/java/com/github/swent/swisstravel/model/trip/RouteSegment.kt
@@ -8,9 +8,7 @@ import kotlin.math.roundToInt
  *
  * @property from The starting location of the route segment.
  * @property to The ending location of the route segment.
- * @property distanceMeter The distance of the route segment in meters.
  * @property durationMinutes The duration of the route segment in minutes.
- * @property path The path of the route segment.
  * @property transportMode The transport mode of the route segment.
  * @property startDate The start date of the route segment.
  * @property endDate The end date of the route segment.
@@ -18,9 +16,7 @@ import kotlin.math.roundToInt
 data class RouteSegment(
     val from: Location,
     val to: Location,
-    val distanceMeter: Int,
     val durationMinutes: Int,
-    val path: List<Coordinate>?,
     val transportMode: TransportMode?,
     val startDate: Timestamp,
     val endDate: Timestamp
@@ -31,12 +27,5 @@ data class RouteSegment(
     val duration = durationMinutes / 60.0
     val roundedDuration = (duration * 100).roundToInt() / 100.0
     return roundedDuration
-  }
-
-  /** Returns the distance in km with two decimals. */
-  fun getDistanceKm(): Double {
-    val distance = distanceMeter / 1000.0
-    val roundedDistance = (distance * 100).roundToInt() / 100.0
-    return roundedDistance
   }
 }
