+++ app/src/main/java/com/github/swent/swisstravel/model/trip/TransportMode.kt
@@ -5,5 +5,6 @@ enum class TransportMode {
   TRAIN,
   BUS,
   TRAM,
+  CAR,
   UNKNOWN
 }
